[Refactor] move interfaces and type definition into `menu.ts`
